requirements.txt
ffmpeg-python>=0.2.0
Pillow>=10.0.0
rich>=13.0.0
tqdm>=4.66.0
requests>=2.31.0
edge-tts>=6.1.10; platform_system!="Linux" or python_version>="3.8"
pyttsx3>=2.90
pydub>=0.25.1
